♻️ extract classes to seperate files and resolve missing documentation warnings
